requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.9
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.3.0
rectools-lightfm==1.17.3
requests==2.31.0
Werkzeug==2.3.7
